requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
Django==4.2
django-range-slider==0.2.4
django-widget-tweaks==1.4.12
exceptiongroup==1.1.1
fake-useragent==1.1.3
filelock==3.11.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
parsel==1.7.0
Pillow==9.5.0
Protego==0.2.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pygame==2.4.0
pyOpenSSL==23.1.1
PySocks==1.7.1
python-vlc==3.0.18122
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
scrapoxy==1.11
Scrapy==2.8.0
scrapy-user-agents==0.1.1
selenium==4.9.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
ua-parser==0.16.1
urllib3==1.26.15
user-agents==2.2.0
w3lib==2.1.1
wget==3.2
wsproto==1.2.0
zope.interface==6.0
